ci: handle gas-oracle latest image, skip builder
This PR updates the `ops/scripts/ci-versions.js` file
to skip building the builder image for when there
are changes to the `gas-oracle`. The `builder` contains
all of the JS code required for the various typescript
based services.

This PR also publishes `ethereumoptimism/gas-oracle:latest`.
It is faster to `docker pull` than `docker build`.
This will speed up the workflow for developers building locally
against the `docker compose` setup as well as the CI if it
is refactored to pull the `latest` images for the unchanged images
instead of building them from scratch.