batch-submitter: remove call to `appendQueueBatch`
The batch submitter is normally ran without running the call to
`appendQueueBatch`. The contracts have `appendQueueBatch` disabled
currently. This means that calling `appendQueueBatch` will result
in a revert. This PR removes the config option that is responsible
for turning of the call to `appendQueueBatch` so it is more likely
to "just work".