l2geth: config enable arbitrary contract deployment
Add a new config option `ROLLUP_ENABLE_ARBITRARY_CONTRACT_DEPLOYMENT`.
This config option must be set with an env var, there is no flag to
set it with.

If it is not set, the functionality is unchanged. If it is set, then it
will override the function `isDeployerAllowed(address)` on the
`OVM_DeployerWhitelist`