contracts: add hardhat impersonate account step
It can be used with the command:

```
$ npx hardhat deploy --tags hardhat --network fork --fork-network kovan
```

A new parameter `--fork-network` is added which specifies the network
that is being forked. We cannot use the actual network due to it
checking the `.chainId` file in the `deployments/xxx/` directory
and comparing that against the remote network.

It is hidden behind the `hardhat` tag and only runs if the deploy config
has the network set to `fork` with `--network fork`. It is up to the
user to configure `hardhat node` with the correct fork url.

This should be used with the env vars:
```
CONTRACTS_TARGET_NETWORK=fork
CONTRACTS_DEPLOYER_KEY=0x...
CONTRACTS_RPC_URL=<url of forked network>
```

This must run first and I didn't want to rename all of the files,
so I named it with the prefix `0000`.