batch-submitter: default tx batch validation to false
Update the batch validation config parsing so
that the tx batch validation is false by default.
This means that to enable transaction batch validation,
the env var `VALIDATE_TX_BATCH` must be explicitly set
to `true`. If it is not set, then it will default to false.
The logic is shown below for the implementation.

```js
> '' ? '' === 'true' : false
false
> 'false' ? 'false' === 'true' : false
false
> 'true' ? 'true' === 'true' : false
true
```