ops: publish `ethereumoptimism/hardhat-node` docker image
Publishing the `hardhat-node` docker image will make using
the `docker-compose` setup much easier for developers
as they will be able to pull all of the docker images needed.
There is an existing `ethereumoptimism/hardhat` image that is
on an older version of `hardhat` that was pushed by the old
Optimism monorepo. This version of hardhat does not support
EIP1559 so it is expected that it will not work with the
existing tech stack.

This PR adds the ability to publish the `hardhat-node` docker
image for both releases and canary releases. A following PR
will be introduced that updates the `docker-compose` files
to specifically reference the new `hardhat-node` image
such that there will not be name collision errors in the future.