feat: add SuggestGasTipCap fallback
This commit adds a fallback in case the L1 provider does not support
queries for eth_maxPriorityFeePerGas. Of our infrastructure providers,
i.e. Alchemy, Infura, Quicknode, and Hardhat (in test envs), only
Alchemy suports this call. In production, we will be routing this call
directly to Alchemy, which should work in the general case. However, if
they were to be unavailable the batch submitter would be stalled until
service was restored. This modification also allows the new EIP-1559
batch submitter to pass our integration tests, since the default Hardhat
backend can't support the query either.