dtl: handle errors in context api
Both `GET /eth/context/latest` and `GET /eth/context/blocknumber/:number`
rely on fetching a remote block to pull the timestamp, blocknumber and
blockhash from. If the fetched block is not found, then properties on
`null` will be attempted to be accessed which will result in a runtime
error. This commit adds handling for this case to prevent that kind of
bug.