gas-oracle: allow configurable base fee poll interval
This commit adds a new config option that can be
configured via argv with `--l1-base-fee-epoch-length-seconds`
or an env var `GAS_PRICE_ORACLE_L1_BASE_FEE_EPOCH_LENGTH_SECONDS`.
It defaults to 15 seconds. Ideally the implementation uses a
duration flag instead of a uint64 flag, but that would be a
breaking change for the config and this service will be
deprecated in the future with the release of bedrock.

By setting this value to a larger number, we are able to update
the L1 base fee that is held in the L2 state less often.
This will save the sequencer money, because it needs to
front the costs of submitting these transactions.