feat: add new contracts-periphery package (#2536)
* feat: add new contracts-periphery package

Introduces a new contracts-periphery package and initializes the package
with the RetroReceiver contract (meant to receive RetroPGF funds on all
available chains). Reason for introducing a new package is that the
original contracts package should be reserved for in-protocol contracts
while contracts-periphery can be used for any contracts that we want to
write and test, regardless of if they're part of the protocol.

* feat(ctp): add periphery contracts to ci

Adds contracts-periphery package to CI in the various places it's
required so we can successfully package and release it.