feat: introduce the drippie-mon service  (#2687)
* feat(ctp): minor Drippie client-side tweaks

Tweaks Drippie to make the client-side a bit simpler. Adds a separate
function "executable" that can be queried to see if a given drip is
currently executable. Updates events so that the name will be properly
emitted but can still be indexed.

* feat: introduce the drippie-mon service

Introduces drippie-mon, a basic service for monitoring Drippie
deployments. drippie-mon will increment a metric whenever a drip is
currently executable. Clients of drippie-mon can watch for this metric
to see if a drip has been executable for a while but has not yet been
executed.