contracts: fix broken contracts tests
The `contracts` tests included a test that did a deep equality check
against two bignumbers. This check stopped working due to changes
in the bignumber implementations used, two bignumbers of the same number
no longer were considered equal. This commit fixes that problem by
using the same version of ethers as `hardhat` and also stringifying
the results so that strings are compared instead of objects. This
fixes the broken tests.