fix(dtl): optimization for mainnet (#3445)
Optimizes the DTL by making the assumption that CTC and SCC addresses in
the AddressManager will NOT change before Bedrock rolls around. This
finally makes it possible to properly sync off of a Geth node where
events are not indexed. You can still sync from Erigon as an alternative
to this commit.