ci: Split docker build & publish
This splits the docker publish job into a build & publish job. They were
previously just steps. The build step uses docker save & circle workspaces
to provide the docker images to dependent jobs (like the publish & hive jobs).

This makes one change to the hive job to remove the docker.pull=true flag. This
is because it loads the image locally but in the case of an external contributor
the image will not be published. Docker will use the local image first & then
pull other images if they are not present.