op-chain-ops: no longer embed receipts
The functionality was added to create a bunch of receipts
as part of the bedrock transition so make the sdk and
withdrawal workflows "just work". We are going to instead
implement logic into the sdk instead of having the extra
receipts. The extra receipts would throw off data people.
It would look like withdrawals happened twice.