op-chain-ops: make the migrated withdrawals version 0
Update the migrated withdrawals to be version 0 based
on the new scheme that is introduced in https://github.com/ethereum-optimism/optimism/pull/4562.
This is a safety measure to ensure that withdrawals cannot
be double withdrawn and that ether stuck in the L1xdm cannot
be stolen in the case of a withdrawal with value reverting.