op-chain-ops: `Withdrawal` type use `hexutil.Bytes`
Use `hexutil.Bytes` in the `Withdrawal` type so that
when writing and reading from JSON, it correctly serializes
the data as a 0x prefixed hex string. Without this, it will
some base64 looking serialization. Be sure to cast to a byte
slice before doing any abi packing, otherwise it will cause
errors.