op-bindings: Check for duplicate artifacts
Checks for duplicate artifacts when instantiating the `hardhat` object. This armors against an issue in the regentool where the regentool was passing in the deployments directory rather than the network directory. This led to artifacts being non-deterministically chosen based on however `fs.WalkDir` iterated over the deployments directory.

Fixes ENG-3054