feat(fd): support Bedrock networks
Updates the fault detector to support Bedrock networks. Bedrock support
is slightly hacky because we still also need to support the legacy
system until mainnet has been upgraded to Bedrock. We will want to
update the fault detector again once mainnet is upgraded to simplify the
service and remove the legacy components.