fix(spec): Add missing call to engine_newPayload
The process to create a new block via the engine API only returns the block and doesn't import it.
So an explicit call to engine_newPayload is required to import the payload before making it canonical.
The actual code performs this call, but it was missed in the spec description.