contracts-bedrock: use kwargs syntax for nonce versioning in tests
Use the named function arguments syntax for encoding the
nonce in the L1CrossDomainMessenger tests. This reduces
the cognitive overhead of ensuring that the tests
are correct. I had to spend the mental energy figuring out
what the arguments were over and over again so i elimated
the problem by moving to the kargs syntax.

I believe that tests should require the least amount of
mental overhead as possible to confirm their correctness.
This is a single commit that reduces the mental overhead
of ensuring our test suite is correct.